[build] Keep workload revision fix focused
Remove the dedicated test project and CI wiring while retaining the normalized revision fix.


Copilot-Session: 70004bfe-9b49-444e-b9be-f15d14383473
